refactor: 공통 폼 스타일 분리
