+++ week4-task/src/pages/MyPage/MyPageSearch.css.ts
@@ -16,10 +16,8 @@ export const resultItem = style({
   padding: "1.2rem",
   marginBottom: "1rem",
   borderRadius: "0.5rem",
-  background: "#f5f6fa",
   fontSize: "1.1rem",
   textAlign: "left",
-  boxShadow: "0 1px 4px rgba(0,0,0,0.03)",
 });
 const spin = keyframes({
   "0%": { transform: "rotate(0deg)" },
@@ -30,7 +28,7 @@ export const spinner = style({
   margin: "2rem 0",
   width: "3rem",
   height: "3rem",
-  border: "0.4rem solid #e0e0e0",
+  border: `0.4rem solid ${vars.color.gray}`,
   borderTop: `0.4rem solid ${vars.color.blue01}`,
   borderRadius: "50%",
   animation: `${spin} 1s linear infinite`,
